add people struct. add endpoint to add employee record. modify people add methods.

+++ modules/hr/src/employees.rs
@@ -46,9 +46,9 @@ impl Employees {
         family_name: &str,
         prefix: &str,
         suffix: &str,
-        gender_id: &i32,
-        ethnicity_id: &i32,
-        marital_state_id: &i32
+        gender_id: &i16,
+        ethnicity_id: &i16,
+        marital_state_id: &i16
     ) -> Result<(), HrError> {
         return Err(HrError::ToBeImplemented(String::from("add")));
     }
